Update UserInterface.java
Adding the functionality to search for recipes based on their ingredients. This is done by choosing the command find ingredient, after which the user is asked for a string. The program should then print all the recipes that contain the specified string. The given string must match exactly the ingredient that is searched for (e.g. "ugar" will return different results than "sugar").